Changed dependency to webgme-engine. Fixes#216 (#215)
* WIP use webgme-engine as dep, tests are running but some failing

* WIP comment out no longer applicable test

* WIP Add back test and fix it

* Bump version to 2.4.1-beta1

* WIP Add webgme as a dev-dep for testing generated files

* WIP added test for missing seed file

* WIP fixed test seed path

* WIP don't remove non-existent dir on failed new seed

* WIP log output/error with build script

* WIP renamed non-existent file

* WIP fixed conflicting seed name

* WIP v0.12.7 support

+++ src/SeedManager.js
@@ -38,13 +38,16 @@ SeedManager.prototype.new = function(args, callback) {
         if (!exists(sourceFile)) {
             var error = 'Seed file does not exist.';
             this._logger.error(error);
-            fse.rmdirSync(fileDir);
             return callback(error);
         }
 
         return fse.copy(sourceFile, filePath, err => {
             if (err) {
-                fse.rmdirSync(fileDir);
+                this._logger.error(err);
+                if (exists(fileDir)) {
+                    fse.rmdirSync(fileDir);
+                }
+                return callback(err);
             }
             this._saveSeed(name, filePath, callback);
         });
@@ -52,7 +55,7 @@ SeedManager.prototype.new = function(args, callback) {
     // This is lazily loaded because the load takes a couple seconds. This
     // causes a delay for calling 'new' but no other commands
     if (!this._exportProject) {
-        var exportCli = require('webgme/src/bin/export');
+        var exportCli = require('webgme-engine/src/bin/export');
         this._exportProject = function (parameters, callback) {
             var args = [
                 'node', 'export.js',
